requirements.txt
# Semi-Intel: Semiconductor Investment Intelligence
# Phase 1 - Data Infrastructure + Phase 2 - Analysis Engine

# Core
fredapi==0.5.2
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
requests>=2.31.0

# Database
# SQLite is built-in; install psycopg2 if migrating to PostgreSQL later
# psycopg2-binary>=2.9.9

# Scheduling
apscheduler>=3.10.0

# Market data
yfinance>=0.2.31

# Analysis (Phase 2)
scikit-learn>=1.3.0    # Regime detection, normalization
# statsmodels>=0.14.0  # Optional: 시계열 분석 고급 기능

# Web API (Phase 3)
fastapi>=0.110.0
uvicorn>=0.27.0
openpyxl>=3.1.0        # Excel export

# AI Advisory (Phase 4) — Multi-Provider
anthropic>=0.40.0         # Anthropic Claude
openai>=1.30.0            # OpenAI GPT
google-genai>=1.0.0       # Google Gemini (기본)

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0          # Pretty console output
loguru>=0.7.0         # Better logging
